[#1674] Chatplugin customize problems (#1678)
* fixed shadow problem

* reset config

* linting
