+++ frontend/chat-plugin/src/components/chat/index.module.scss
@@ -1,3 +1,5 @@
+@import 'assets/scss/fonts.scss';
+
 .main {
   height: 100%;
   padding: 0;
@@ -84,7 +86,7 @@
   position: relative;
 }
 
-.disconnectedOverlay {
+.modalOverlay {
   position: absolute;
   top: 0;
   left: 0;
@@ -95,6 +97,21 @@
   justify-content: center;
   align-items: center;
   color: white;
+}
+
+.modalCloseChat {
+  @include font-m;
+  width: 100%;
+  margin: 16px;
+  padding: 32px 0 38px 0;
+  background: white;
+  border-radius: 8px;
+  p {
+    text-align: center;
+    font-size: 18px;
+    margin: 0 0 8px 0;
+    color: gray;
+  }
 }
 
 .buttonWrapper {
@@ -114,7 +131,7 @@
   height: 40px;
   background-color: white;
   color: var(--color-airy-blue);
-  border-radius: 4px;
+  border-radius: 8px;
   border: 1px solid var(--color-airy-blue);
   cursor: pointer;
   padding: 8px 10px;
@@ -130,7 +147,7 @@
   height: 40px;
   background-color: var(--color-airy-blue);
   color: white;
-  border-radius: 4px;
+  border-radius: 8px;
   border: none;
   cursor: pointer;
   padding: 8px 10px;
